requirements.txt
# Core Dependencies for RAG System

# LlamaIndex for indexing and document management
llama-index==0.12.16  # Update based on the latest version

# ChromaDB for efficient vector-based document retrieval
chromadb==0.3.29  # Update based on the latest version

# OpenAI's GPT-4 for response generation
openai==1.61.1  # Latest version for GPT-4

# DeepEval for response evaluation
deepeval==2.3.6  # Update based on the latest version

# Ragas for relevance-based evaluation
ragas==0.2.13  # Update based on the latest version

# TrueLens for coherence evaluation
trulens==1.3.5  # Update based on the latest version

# Galileo for user feedback evaluation
galileo==0.0.1  # Update based on the latest version

# Utilities

# For text processing and NLP
nltk==3.9.1
spacy==3.8.4

# For handling CSV, JSON, and other data formats
pandas==2.2.3
numpy==2.2.2

# For efficient caching
redis==5.2.1

# For logging and monitoring
loguru==0.7.3

# For file handling and OS utilities
python-dotenv==1.0.1  # To manage environment variables

# Web scraping (optional if ingestion involves web scraping)
beautifulsoup4==4.13.3
requests==2.26.0

# For database interactions (optional if using databases like MongoDB, SQL)
sqlalchemy==2.0.38
pymongo==3.12.1

# For parallel processing and async handling
celery==5.4.0
asyncio==3.4.3  # Python's built-in async library, may be required for async tasks

# For testing and validation
pytest==8.3.4

# For environment management (optional but recommended)
virtualenv==20.10.0

# For deploying and containerization
docker==7.1.0
